Update to 0.8.2
Update to 0.8.2

+++ src/Database/Schema/MysqlSchema.php
@@ -25,7 +25,7 @@ class MysqlSchema extends Schema {
 	}
 
 	public function double($column, $length = [11, 2], $nullable = false, $default = '') {
-		return $this->addColumn($column, 'tinyint', $length, $nullable, $default);
+		return $this->addColumn($column, 'float', $length, $nullable, $default);
 	}
 
 	public function toSQL() {
